Feat(sensor): add factory methods for sensors generation

+++ src/main/java/frc/lib/generic/sensors/Sensor.java
@@ -32,6 +32,6 @@ public class Sensor implements LoggableHardware {
         public int currentValue = 0;
 
         public int[] threadCurrentValue = new int[0];
-        public int[] timestamp = new int[0];
+        public double[] timestamp = new double[0];
     }
 }
